Refactor face extraction to its own class
- should also fix duplicate adding of the same face
